+++ kuStudy/kuStudyKit/Library.swift
@@ -53,7 +53,7 @@ public class LibraryViewModel {
         available = library.available
         
         let infoRealm = kuStudy.infoRealm()
-        let record = infoRealm.objectForPrimaryKey(LibraryInfoRecord.self, key: library.id)
+        let record = infoRealm?.objectForPrimaryKey(LibraryInfoRecord.self, key: library.id)
         key = record?.key != nil ? record!.key : "unknown_id"
         name = NSLocalizedString(key, bundle: NSBundle(forClass: self.dynamicType), comment: "Library key") // TODO: Use localized string
     }