+++ Release Notes.md
@@ -1,3 +1,9 @@
 # Release Notes
 
 -----
+
+## 0.1.0 (1)
+
+* iPhone app (8.3 and above)
+* Today extension
+* Apple Watch app with Glance